chore(clippy): fix 12 workspace lint errors under -D warnings
- 6× inconsistent_digit_grouping: regroup literal underscores in
  uniswap_v2.rs, uniswap_v3.rs, common/types.rs
- 4× unnecessary_cast: drop `as u64` on already-u64 base_fee_per_gas in
  three integration tests
- collapsible_match + single_match at anvil_fork_test.rs: fold the outer
  `if let Some(event)` and inner single-arm `match` into one `if let`
  destructuring pattern
- manual_range_contains at bellman_ford.rs:644: `v >= 3 && v <= 5`
  becomes `(3..=5).contains(&v)`
- unnecessary_map_or at fork.rs:328: switch to `is_none_or`
- dead_code at simulator/lib.rs: delete unused `setup_state_with_caller`
  helper

+++ crates/integration-tests/tests/mainnet_fork_flashloan_test.rs
@@ -87,7 +87,7 @@ async fn run_flash_loan_test(anvil_url: &str) -> Result<(), String> {
         .ok_or("block not found")?;
 
     let timestamp = block.header.timestamp;
-    let base_fee = block.header.base_fee_per_gas.unwrap_or(30_000_000_000) as u64;
+    let base_fee = block.header.base_fee_per_gas.unwrap_or(30_000_000_000);
 
     phase_times.push(("Phase 1 (provider setup)", t_phase1.elapsed().as_millis()));
     eprintln!("\n=== Flash Loan Arbitrage E2E Test ===");
@@ -435,7 +435,7 @@ async fn run_flash_loan_test(anvil_url: &str) -> Result<(), String> {
         .ok_or("block not found for sim")?;
     let current_ts = current_block_data.header.timestamp;
     let current_base_fee =
-        current_block_data.header.base_fee_per_gas.unwrap_or(30_000_000_000) as u64;
+        current_block_data.header.base_fee_per_gas.unwrap_or(30_000_000_000);
 
     let sim_parsed: url::Url = anvil_url.parse().expect("valid URL");
     let sim_provider = ProviderBuilder::new().connect_http(sim_parsed);
